Directory class wasn't masked; as a result, using it caused crash. As a fix, we removed DirectoryOpen function and allowed Directory constructor to work as previously done by DirectoryOpen.
git-svn-id: svn://falconpl.org/falcon/core/trunk@1056 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/core_module/functional_ext.cpp
@@ -2081,7 +2081,7 @@ static bool core_brigade_next( ::Falcon::VMachine *vm )
    if ( vm->local(1)->isArray() )
    {
       CoreArray* callarr = vm->local(1)->asArray();
-      callarr->at(0) = list->at(next);
+      callarr->at(0) = list->at((int32)next);
       success = vm->callFrame( callarr, 0 );
    }
    else 
@@ -2092,7 +2092,7 @@ static bool core_brigade_next( ::Falcon::VMachine *vm )
          vm->pushParameter( *vm->param(i) );
       }
       
-      success = vm->callFrame( list->at(next), vm->paramCount()-1 );
+      success = vm->callFrame( list->at((int32)next), vm->paramCount()-1 );
    }
    
    if ( ! success )
